tools/modem_vuln_scanner.py
import nmap

def scan_modem(ip_address):
    """
    Scans a network modem for common vulnerabilities.
    """
    print(f"[*] Scanning modem at {ip_address} for vulnerabilities...")
    nm = nmap.PortScanner()
    # This is a placeholder for a real implementation.
    # In a real-world scenario, you would use a more comprehensive set of scripts and techniques.
    nm.scan(ip_address, arguments='-sV --script=vuln')
    return nm.csv()

if __name__ == "__main__":
    import argparse
    parser = argparse.ArgumentParser(description="Network Modem Vulnerability Scanner")
    parser.add_argument("ip_address", help="The IP address of the modem to scan.")
    args = parser.parse_args()
    results = scan_modem(args.ip_address)
    print(results)
