requirements.txt
# Core dependencies
requests>=2.28.0
colorama>=0.4.4
beautifulsoup4>=4.11.0
urllib3>=1.26.0

# DNS and network utilities
dnspython>=2.2.0
python-nmap>=0.7.1
netaddr>=0.8.0
netifaces>=0.11.0

# Web scraping and parsing
lxml>=4.9.0
selenium>=4.8.0

# Security and cryptography
cryptography>=3.4.8
pyopenssl>=22.0.0

# Data processing and analysis
pandas>=1.5.3
numpy>=1.24.2
python-dateutil>=2.8.2

# Configuration and serialization
pyyaml>=6.0
jinja2>=3.1.2

# CLI and output formatting
click>=8.1.3
rich>=13.3.1
tabulate>=0.9.0
termcolor>=2.2.0
tqdm>=4.64.1

# System utilities
psutil>=5.9.4

# Optional API integrations (install as needed)
shodan>=1.28.0
pygithub>=1.55
censys>=2.1.9
virustotal-api>=1.1.11
sherlock-project>=0.14.3
